requirements.txt
requests
finnhub-python
python-dotenv
yfinance
cerebras_cloud_sdk
matplotlib
fastapi
uvicorn
